challenge 1 -  create a function similar to linespace, just use 'arange' instead.
Write the code above using the NumPy function arange() instead of linspace(). If you need to, read the documentation for these functions.

The output is an array with the start number only.
As the breakdown range specified in the function is 5000. Where as, the upper limit of the array is 100.0 only.
Hence, it only creates one element in the array.